feat: refatoração da program.cs
