+++ src/Venice.Orders.WebApi/Program.cs
@@ -15,6 +15,7 @@ using Venice.Orders.Application.Interfaces;
 using Venice.Orders.Application;
 using Venice.Orders.WebApi.Middleware;
 using Venice.Orders.WebApi.HealthChecks;
+using Venice.Orders.WebApi.Configuration;
 
 var builder = WebApplication.CreateBuilder(args);
 
@@ -31,320 +32,20 @@ MongoDbConfigurationService.ConfigureSerialization();
 builder.Services.AddControllers();
 builder.Services.AddEndpointsApiExplorer();
 
-// Add CORS
-builder.Services.AddCors(options =>
-{
-    options.AddPolicy("AllowAll", policy =>
-    {
-        var allowedOrigins = builder.Configuration.GetSection("Cors:AllowedOrigins").Get<string[]>() ?? new[] { "*" };
-        var allowedMethods = builder.Configuration.GetSection("Cors:AllowedMethods").Get<string[]>() ?? new[] { "*" };
-        var allowedHeaders = builder.Configuration.GetSection("Cors:AllowedHeaders").Get<string[]>() ?? new[] { "*" };
-
-        if (allowedOrigins.Contains("*"))
-        {
-            policy.AllowAnyOrigin();
-        }
-        else
-        {
-            policy.WithOrigins(allowedOrigins);
-        }
-
-        if (allowedMethods.Contains("*"))
-        {
-            policy.WithMethods(allowedMethods);
-        }
-        else
-        {
-            policy.AllowAnyMethod();
-        }
-
-        if (allowedHeaders.Contains("*"))
-        {
-            policy.AllowAnyHeader();
-        }
-        else
-        {
-            policy.WithHeaders(allowedHeaders);
-        }
-    });
-});
-builder.Services.AddSwaggerGen(c =>
-{
-    c.SwaggerDoc("v1", new Microsoft.OpenApi.Models.OpenApiInfo
-    {
-        Title = "Venice Orders API",
-        Version = "v1",
-        Description = "API for managing orders in the Venice system"
-    });
-
-    // Include XML comments for better documentation
-    var xmlFile = $"{System.Reflection.Assembly.GetExecutingAssembly().GetName().Name}.xml";
-    var xmlPath = Path.Combine(AppContext.BaseDirectory, xmlFile);
-    if (File.Exists(xmlPath))
-    {
-        c.IncludeXmlComments(xmlPath);
-    }
-
-    // Custom schema ID resolver to avoid conflicts
-    c.CustomSchemaIds(type =>
-    {
-        if (type.Name == "OrderItemResult")
-        {
-            return $"{type.Namespace?.Split('.').Last()}.{type.Name}";
-        }
-        return type.Name;
-    });
-
-    // Add JWT authentication support
-    c.AddSecurityDefinition("Bearer", new Microsoft.OpenApi.Models.OpenApiSecurityScheme
-    {
-        Description = "JWT Authorization header using the Bearer scheme. Example: \"Authorization: Bearer {token}\"",
-        Name = "Authorization",
-        In = Microsoft.OpenApi.Models.ParameterLocation.Header,
-        Type = Microsoft.OpenApi.Models.SecuritySchemeType.ApiKey,
-        Scheme = "Bearer"
-    });
-
-    c.AddSecurityRequirement(new Microsoft.OpenApi.Models.OpenApiSecurityRequirement
-    {
-        {
-            new Microsoft.OpenApi.Models.OpenApiSecurityScheme
-            {
-                Reference = new Microsoft.OpenApi.Models.OpenApiReference
-                {
-                    Type = Microsoft.OpenApi.Models.ReferenceType.SecurityScheme,
-                    Id = "Bearer"
-                }
-            },
-            Array.Empty<string>()
-        }
-    });
-});
-
-// JWT Authentication
-builder.Services.AddAuthentication(JwtBearerDefaults.AuthenticationScheme)
-    .AddJwtBearer(options =>
-    {
-        options.TokenValidationParameters = new TokenValidationParameters
-        {
-            ValidateIssuer = true,
-            ValidateAudience = true,
-            ValidateLifetime = true,
-            ValidateIssuerSigningKey = true,
-            ValidIssuer = builder.Configuration["Jwt:Issuer"] ?? "VeniceOrders",
-            ValidAudience = builder.Configuration["Jwt:Audience"] ?? "VeniceOrders",
-            IssuerSigningKey = new SymmetricSecurityKey(
-                Encoding.UTF8.GetBytes(builder.Configuration["Jwt:Key"] ?? "your-secret-key-here"))
-        };
-    });
-
-builder.Services.AddAuthorization();
-
-// Entity Framework - SQL Server
-builder.Services.AddDbContext<VeniceOrdersContext>(options =>
-    options.UseSqlServer(builder.Configuration.GetConnectionString("SqlServer")));
-
-// MongoDB
-builder.Services.AddSingleton<IMongoClient>(provider =>
-{
-    var connectionString = builder.Configuration.GetConnectionString("MongoDB");
-    
-    // Configure MongoDB settings with proper Guid representation
-    var settings = MongoClientSettings.FromConnectionString(connectionString);
-    
-    return new MongoClient(settings);
-});
-
-builder.Services.AddSingleton<IMongoDatabase>(provider =>
-{
-    var client = provider.GetRequiredService<IMongoClient>();
-    var databaseName = builder.Configuration["MongoDB:DatabaseName"] ?? "VeniceOrders";
-    return client.GetDatabase(databaseName);
-});
-
-// Redis
-builder.Services.AddSingleton<IConnectionMultiplexer>(provider =>
-{
-    var connectionString = builder.Configuration.GetConnectionString("Redis");
-    if (string.IsNullOrEmpty(connectionString))
-    {
-        throw new InvalidOperationException("Redis connection string is not configured");
-    }
-    return ConnectionMultiplexer.Connect(connectionString);
-});
-
-// RabbitMQ
-builder.Services.AddSingleton<IConnection>(provider =>
-{
-    var connectionString = builder.Configuration.GetConnectionString("RabbitMQ");
-    if (string.IsNullOrEmpty(connectionString))
-    {
-        throw new InvalidOperationException("RabbitMQ connection string is not configured");
-    }
-    var factory = new ConnectionFactory { Uri = new Uri(connectionString) };
-    return factory.CreateConnection();
-});
-
-// Repositories
-builder.Services.AddScoped<IOrderRepository, OrderRepository>();
-builder.Services.AddScoped<IOrderItemRepository, OrderItemRepository>();
-builder.Services.AddScoped<IUserRepository, UserRepository>();
-
-// Services
-builder.Services.AddScoped<ICacheService, RedisCacheService>();
-builder.Services.AddSingleton<IEventPublisher, RabbitMQEventPublisher>();
-
-
-
-// MediatR
-builder.Services.AddMediatR(cfg => cfg.RegisterServicesFromAssembly(typeof(ApplicationLayer).Assembly));
-
-// AutoMapper
-builder.Services.AddAutoMapper(typeof(Program).Assembly, typeof(ApplicationLayer).Assembly);
-
-// Health Checks
-builder.Services.AddHealthChecks()
-    .AddCheck<CustomHealthCheck>("application", tags: new[] { "app" })
-    .AddCheck<ExternalServicesHealthCheck>("external_services", tags: new[] { "external" })
-    .AddSqlServer(
-        builder.Configuration.GetConnectionString("SqlServer") ?? string.Empty,
-        name: "sqlserver",
-        tags: new[] { "database", "sql" },
-        timeout: TimeSpan.FromSeconds(5))
-    .AddMongoDb(
-        provider => 
-        {
-            var connectionString = builder.Configuration.GetConnectionString("MongoDB") ?? string.Empty;
-            var settings = MongoClientSettings.FromConnectionString(connectionString);
-            return new MongoClient(settings);
-        },
-        name: "mongodb",
-        tags: new[] { "database", "nosql" },
-        timeout: TimeSpan.FromSeconds(5))
-    .AddRedis(
-        builder.Configuration.GetConnectionString("Redis") ?? string.Empty,
-        name: "redis",
-        tags: new[] { "cache" },
-        timeout: TimeSpan.FromSeconds(5))
-    .AddDbContextCheck<VeniceOrdersContext>(
-        name: "ef_core",
-        tags: new[] { "database", "ef" });
-
-
+// Configure application services
+builder.Services
+    .AddCorsServices(builder.Configuration)
+    .AddSwaggerServices()
+    .AddAuthenticationServices(builder.Configuration)
+    .AddApplicationServices(builder.Configuration)
+    .AddApplicationHealthChecks(builder.Configuration);
 
 var app = builder.Build();
 
-// Configure the HTTP request pipeline.
-// Enable Swagger for both Development and Docker environments
-if (app.Environment.IsDevelopment() || app.Environment.EnvironmentName == "Docker")
-{
-    app.UseSwagger();
-    app.UseSwaggerUI(c =>
-    {
-        c.SwaggerEndpoint("/swagger/v1/swagger.json", "Venice Orders API v1");
-        // Swagger UI will be available at /swagger
-    });
-}
+// Configure the HTTP request pipeline
+app.UseApplicationPipeline(builder.Environment);
 
-// Add middleware
-app.UseMiddleware<ValidationExceptionMiddleware>();
-
-// Use CORS
-app.UseCors("AllowAll");
-
-// HTTPS redirection disabled for Docker environment
-
-app.UseAuthentication();
-app.UseAuthorization();
-
-app.MapControllers();
-
-// Health Check endpoints
-app.MapHealthChecks("/health", new Microsoft.AspNetCore.Diagnostics.HealthChecks.HealthCheckOptions
-{
-    ResponseWriter = async (context, report) =>
-    {
-        context.Response.ContentType = "application/json";
-
-        var result = new
-        {
-            status = report.Status.ToString(),
-            timestamp = DateTime.UtcNow,
-            duration = report.TotalDuration,
-            checks = report.Entries.Select(e => new
-            {
-                name = e.Key,
-                status = e.Value.Status.ToString(),
-                description = e.Value.Description,
-                duration = e.Value.Duration,
-                tags = e.Value.Tags,
-                data = e.Value.Data
-            })
-        };
-
-        await context.Response.WriteAsJsonAsync(result);
-    }
-});
-
-app.MapHealthChecks("/health/ready", new Microsoft.AspNetCore.Diagnostics.HealthChecks.HealthCheckOptions
-{
-    Predicate = check => check.Tags.Contains("external") || check.Tags.Contains("database"),
-    ResponseWriter = async (context, report) =>
-    {
-        context.Response.ContentType = "application/json";
-
-        var result = new
-        {
-            status = report.Status.ToString(),
-            timestamp = DateTime.UtcNow,
-            duration = report.TotalDuration,
-            checks = report.Entries.Select(e => new
-            {
-                name = e.Key,
-                status = e.Value.Status.ToString(),
-                description = e.Value.Description,
-                duration = e.Value.Duration
-            })
-        };
-
-        await context.Response.WriteAsJsonAsync(result);
-    }
-});
-
-app.MapHealthChecks("/health/live", new Microsoft.AspNetCore.Diagnostics.HealthChecks.HealthCheckOptions
-{
-    Predicate = check => check.Tags.Contains("app"),
-    ResponseWriter = async (context, report) =>
-    {
-        context.Response.ContentType = "application/json";
-
-        var result = new
-        {
-            status = report.Status.ToString(),
-            timestamp = DateTime.UtcNow,
-            duration = report.TotalDuration
-        };
-
-        await context.Response.WriteAsJsonAsync(result);
-    }
-});
-
-// Apply migrations when running in Docker
-if (app.Environment.EnvironmentName == "Docker")
-{
-    using (var scope = app.Services.CreateScope())
-    {
-        try
-        {
-            var context = scope.ServiceProvider.GetRequiredService<VeniceOrdersContext>();
-            context.Database.Migrate();
-            Console.WriteLine("Database migrations applied successfully");
-        }
-        catch (Exception ex)
-        {
-            Console.WriteLine($"Error applying migrations: {ex.Message}");
-        }
-    }
-}
+// Apply database migrations
+app.ApplyDatabaseMigrations(builder.Environment);
 
 app.Run();
